Fix C++ bug
